Добавил фз 10

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -2,13 +2,11 @@ package ru.yandex.practicum.filmorate.model;
 
 
 import lombok.Data;
-
-import javax.validation.constraints.*;
 import java.time.LocalDate;
 
 @Data
 public class Film {
-    private int id;
+    private Long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
