+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -2,11 +2,12 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.Data;
 import java.time.LocalDate;
+import java.util.HashSet;
+import java.util.Set;
 
 @Data
 public class User {
-
-    private int id;
+    private Long id;
     private String email;
     private String login;
     private String name;
